🎨 sample페이지 생성하여 분리하고 메인페이지 따로 생성
